refactor: move tile types to separate file
