+++ bciers/apps/dashboard/app/dashboard/page.tsx
@@ -1,6 +1,7 @@
-import { Tiles } from "@bciers/components/server";
+import Tiles from "@bciers/components/navigation/Tiles";
+import Note from "@bciers/components/layout/Note";
 import { fetchDashboardData } from "@bciers/actions/server";
-import { ContentItem } from "@bciers/types";
+import { ContentItem } from "@bciers/types/tiles";
 
 export default async function Page() {
   // 🚀 API fetch dashboard tiles
@@ -12,6 +13,11 @@ export default async function Page() {
   // Build the navigation tiles
   return (
     <>
+      <Note variant="important">
+        <b>Important:</b> Please always ensure that the information in{" "}
+        <b>Registration</b> is complete and accurate before submit or amend
+        reports in <b>Reporting.</b>
+      </Note>
       <Tiles tiles={data} />
     </>
   );
