+++ bciers/libs/shared/components/src/lib/layout/Main.tsx
@@ -4,7 +4,7 @@ type MainProps = {
 
 const Main = ({ children }: MainProps) => {
   return (
-    <main className="w-full max-w-[1536px] min-h-full mx-auto p-4 sm:p-6 pb-48 sm:pb-20">
+    <main className="w-full max-w-[1536px] min-h-full mx-auto px-4 sm:px-6 pb-48 sm:pb-20">
       {children}
     </main>
   );
