+++ bciers/tsconfig.base.json
@@ -29,8 +29,8 @@
       "@bciers/styles": ["libs/shared/styles/src/index.ts"],
       "@bciers/styles/*": ["libs/shared/styles/src/lib/*"],
       "@bciers/testConfig/*": ["libs/shared/testConfig/src/*"],
-      "@bciers/types": ["libs/shared/types/src/index.ts"],
-      "@bciers/utils/*": ["libs/shared/utils/*"]
+      "@bciers/utils/*": ["libs/shared/utils/*"],
+      "@bciers/types/*": ["libs/shared/types/src/*"]
     }
   },
   "include": ["**/*.ts", "**/*.tsx", "**/*.js", "**/*.jsx", "**/*.json"],
